refactor(createAsset): ♻️ moved components to partials
