+++ components/partials/create_asset/ControlWindow.tsx
